Use consistent name for the "get item from untyped map iterator" thunk
It now uses the same prefix as other thunks needed for the proxied type,
so the RawMapThunk helper can be used for enums.

Calling it a "iter next" thunk is misleading.
It does not increment the iterator as "next" implies,
it only gets the current key/value the iterator points to.

PiperOrigin-RevId: 608800860